refactor(api): move and rework archive-organization tests
