+++ api/tests/unit/application/organizations/index_test.js
@@ -109,32 +109,4 @@ describe('Unit | Router | organization-router', function () {
       expect(response.statusCode).to.equal(403);
     });
   });
-
-  describe('POST /api/admin/organizations/{id}/archive', function () {
-    it('returns forbidden access if admin member has CERTIF role', async function () {
-      // given
-      sinon.stub(organizationController, 'archiveOrganization').resolves('ok');
-
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('POST', '/api/admin/organizations/1/archive');
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      sinon.assert.notCalled(organizationController.archiveOrganization);
-    });
-  });
 });
